Moves OperatorList and QueueOptimizer into separate file.
