+++ test/unit/evaluator_spec.js
@@ -15,8 +15,9 @@
 
 import { Dict, Name } from '../../src/core/primitives';
 import { FormatError, OPS } from '../../src/shared/util';
-import { OperatorList, PartialEvaluator } from '../../src/core/evaluator';
 import { Stream, StringStream } from '../../src/core/stream';
+import { OperatorList } from '../../src/core/operator_list';
+import { PartialEvaluator } from '../../src/core/evaluator';
 import { WorkerTask } from '../../src/core/worker';
 import { XRefMock } from './test_utils';
